+++ shiori-backend/src/modules/integration/integration.controller.js
@@ -1,0 +1,10 @@
+import * as integrationService from "./integration.service.js";
+
+export const getIntegrationStatus = (req, res, next) => {
+  try {
+    const status = integrationService.getIntegrationStatus();
+    res.json(status);
+  } catch (error) {
+    next(error);
+  }
+};